Fixes odd scrolling bug
When the JS sees you've scrolled to a certain number of pixels, it moves the header margin to an absolute position on the screen thus removing the margin that is added by the code. By changing the value to height, the JS checks for the height of the header and adds that to the margin-top of the maincontent class.